* game + fix warning: text set.CharacterSize(float) -> int * game state + fix warning: EGameState static -> non static * MainMenuGS + fix warning: implicit cast size_t to uint -> explicit cast

+++ LineBreaker/MainMenuGS.cpp
@@ -59,7 +59,7 @@ void MainMenuGS::handleEvent(sf::Event evnt)
             break;
         case sf::Keyboard::W:           // move one option up, loop from top to bottom
         case sf::Keyboard::Up:
-            selectedItem = (selectedItem == 0) ? (items.size() - 1) : (selectedItem - 1);
+            selectedItem = (selectedItem == 0) ? static_cast<unsigned int>(items.size() - 1) : (selectedItem - 1);
             break;
         default:
             break;
